Gather various functions into `io_KinDA` module
