+++ src/kinda/__init__.py
@@ -8,7 +8,8 @@ from importlib import metadata
 __version__ = f"v{metadata.version('kinda')}"
 
 
-from .kinda import System, from_pil, import_data, export_data
+from .kinda import System
+from .objects.io_KinDA import import_data, export_data, from_pil
 from . import statistics
 from . import simulation
 from . import enumeration
